requirements.txt
biom-format==2.1.16
biopython==1.85
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
fonttools==4.55.3
h5py==3.12.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
natsort==8.4.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyVCF3==1.0.3
requests==2.32.3
scikit-bio==0.6.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
statsmodels==0.14.4
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
